modules/data_handler/requirements.txt
azure-storage-blob
flask
psycopg2
pytz
gunicorn